% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2.R
\name{doPermute_seq_correlation}
\alias{doPermute_seq_correlation}
\title{Do permutation for calculate score}
\usage{
doPermute_seq_correlation(
  sets,
  seuratObj,
  genes_selected,
  type1,
  type2,
  returnType = "network",
  seedPerm = 0,
  scoreMethod = "sum"
)
}
\arguments{
\item{setInd}{list of genesetsets}
}
\description{
Do permutation for calculate score
}
