% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2.R
\name{getPvalues}
\alias{getPvalues}
\title{Calculate p values with permutation}
\usage{
getPvalues(cor_data_score, cor_data_perm_list)
}
\arguments{
\item{cor_data_score}{score of pathways between two groups}

\item{cor_data_perm_list}{permutation results}
}
\value{
A numeric vector named by pathways
}
\description{
Calculate p values with permutation
}
