% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2.R
\name{getPvaluesTTest}
\alias{getPvaluesTTest}
\title{Calculate p values with permutation}
\usage{
getPvaluesTTest(cor_data_score, cor_data_perm_list)
}
\arguments{
\item{cor_data_score}{score of pathways between two groups}

\item{cor_data_perm_list}{permutation results}
}
\description{
Calculate p values with permutation
}
