% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2.R
\name{GSNCAscore}
\alias{GSNCAscore}
\title{compute the score with GSCNA approach}
\usage{
GSNCAscore(setInd, twonetworks)
}
\arguments{
\item{setInd}{list of geneset}

\item{twonetworks}{list of two networks}
}
\value{
numeric value
}
\description{
compute the score with GSCNA approach
}
