% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2.R
\name{network_plot}
\alias{network_plot}
\title{Do permutation for calculate score}
\usage{
network_plot(
  networklist,
  name1 = "status1",
  name2 = "status2",
  fileName = NULL
)
}
\arguments{
\item{networklist}{list with two networks}
}
\description{
Do permutation for calculate score
}
