% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2.R
\name{runPermute_seq_correlation}
\alias{runPermute_seq_correlation}
\title{Run permutation for calculate score}
\usage{
runPermute_seq_correlation(
  sets,
  seuratObj,
  genes_selected,
  type1,
  type2,
  ncores = 1,
  nPermute = 6,
  returnType = "network",
  scoreMethod = "sum"
)
}
\arguments{
\item{seuratObj}{A Seurat Oject, with status in meta.data}

\item{genes_selected}{Gene selected to be analyzed}

\item{type1}{the first type (defined in status) for analysis}

\item{type2}{the second type (defined in status) for analysis}

\item{returnType}{return network or network score}

\item{scoreMethod}{method of score pathway difference (sum or GSNCA)}

\item{setInd}{list of genesetsets}

\item{seedPerm}{seed for permutaion}
}
\value{
list of list of pathway scores
}
\description{
Run permutation for calculate score
}
