% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2.R
\name{setsCorrelationScore}
\alias{setsCorrelationScore}
\title{compute set correlation score with SUM method}
\usage{
setsCorrelationScore(setInd, corDiff)
}
\arguments{
\item{setInd}{list of geneset}

\item{corDiff}{matrix of differential correlation}
}
\value{
numeric value
}
\description{
compute set correlation score with SUM method
}
