% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2.R
\name{calculate_seq_correlation}
\alias{calculate_seq_correlation}
\title{calculate score of list of gene sets}
\usage{
calculate_seq_correlation(
  sets,
  seuratObj,
  genes_selected,
  type1,
  type2,
  returnType = "network",
  scoreMethod = "sum"
)
}
\arguments{
\item{setInd}{list of genesetsets}
}
\description{
calculate score of list of gene sets
}
