% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2.R
\name{network_seq_correlation}
\alias{network_seq_correlation}
\title{calculate the network of gene sets}
\usage{
network_seq_correlation(
  pathway,
  seuratObj,
  genes_selected,
  type1,
  type2,
  returnType = "network"
)
}
\arguments{
\item{pathway}{a gene list}
}
\description{
calculate the network of gene sets
}
