% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2.R
\name{runPermute_seq_correlation}
\alias{runPermute_seq_correlation}
\title{Run permutation for calculate score}
\usage{
runPermute_seq_correlation(
  sets,
  seuratObj,
  genes_selected,
  type1,
  type2,
  ncores = 1,
  nPermute = 6,
  returnType = "network",
  scoreMethod = "sum"
)
}
\arguments{
\item{setInd}{list of genesetsets}
}
\description{
Run permutation for calculate score
}
